Связующее звено
