+++ frontend/controllers/TasksController.php
@@ -245,30 +245,32 @@ class TasksController extends SecuredController
             }
         }
 
-        $messagesForm = new Messages();
 
-        if (\Yii::$app->request->getIsPost()) {
-            $messagesForm->load(\Yii::$app->request->post());
 
-            if ($messagesForm->validate()) {
-                if($messagesForm->save()) {
+        // $messagesForm = new Messages();
 
-                    $email->messageAction();
+        // if (\Yii::$app->request->getIsPost()) {
+        //     $messagesForm->load(\Yii::$app->request->post());
 
-                    $this->redirect(['tasks/view', 'id' => \Yii::$app->request->get('id')]);
-                }
-            }
-        }
+        //     if ($messagesForm->validate()) {
+        //         if($messagesForm->save()) {
 
-        $messages = Messages::find()
-            ->where(['task_id' => $id])
-            ->orderBy('date_add DESC')
-            ->limit(1)
-            ->all();
+        //             $email->messageAction();
 
-        $messages = Messages::find()
-            ->where(['task_id' => $id])
-            ->all();
+        //             $this->redirect(['tasks/view', 'id' => \Yii::$app->request->get('id')]);
+        //         }
+        //     }
+        // }
+
+        // $messages = Messages::find()
+        //     ->where(['task_id' => $id])
+        //     ->orderBy('date_add DESC')
+        //     ->limit(1)
+        //     ->all();
+
+        // $messages = Messages::find()
+        //     ->where(['task_id' => $id])
+        //     ->all();
 
         return $this->render(
             'view',
@@ -285,8 +287,8 @@ class TasksController extends SecuredController
                 'creatorID',
                 'oneRespond',
                 'user',
-                'messagesForm',
-                'messages'
+                // 'messagesForm',
+                // 'messages'
             )
         );
     }
