+++ frontend/models/Messages.php
@@ -33,14 +33,14 @@ class Messages extends \yii\db\ActiveRecord
      */
     public function rules()
     {
-        $creatorID = null;
-        $executorID = null;
+        // $creatorID = null;
+        // $executorID = null;
 
-        if (\Yii::$app->user->getId() === Yii::$app->params['task_current']->user_id_create) {
-            $creatorID = Yii::$app->params['task_current']->user_id_create;
-        } else {
-            $executorID = Yii::$app->user->getId();
-        }
+        // if (\Yii::$app->user->getId() === Yii::$app->params['task_current']->user_id_create) {
+        //     $creatorID = Yii::$app->params['task_current']->user_id_create;
+        // } else {
+        //     $executorID = Yii::$app->user->getId();
+        // }
 
         return [
             [['message'], 'required'],
@@ -54,9 +54,9 @@ class Messages extends \yii\db\ActiveRecord
             [['task_id'], 'exist', 'skipOnError' => true, 'targetClass' => Tasks::class,
             'targetAttribute' => ['task_id' => 'id']],
             ['date_add', 'default', 'value' => Yii::$app->formatter->asDate('now', 'yyyy-MM-dd H:m:s')],
-            ['task_id', 'default', 'value' => Yii::$app->params['task_current']->id],
-            ['user_id_executor', 'default', 'value' => $executorID],
-            ['user_id_create', 'default', 'value' => $creatorID],
+            // ['task_id', 'default', 'value' => Yii::$app->params['task_current']->id],
+            // ['user_id_executor', 'default', 'value' => $executorID],
+            // ['user_id_create', 'default', 'value' => $creatorID],
         ];
     }
