+++ frontend/modules/api/Module.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace frontend\modules\api;
+
+use yii\filters\AccessControl;
+
+/**
+ * api module definition class
+ */
+class Module extends \yii\base\Module
+{
+    public function behaviors()
+    {
+        return [
+            'access' => [
+                'class' => AccessControl::class,
+                'rules' => [
+                    [
+                        'allow' => true,
+                        'roles' => ['@']
+                    ]
+                ]
+            ]
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public $controllerNamespace = 'frontend\modules\api\controllers';
+
+    /**
+     * {@inheritdoc}
+     */
+    public function init()
+    {
+        parent::init();
+
+        // custom initialization code goes here
+    }
+}
